backend/pyproject.toml
[project]
name = "whealth-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.8.2",
  "sqlalchemy==2.0.32",
  "psycopg[binary]>=3.2.2,<3.3",
  "alembic==1.13.2",
  "python-jose==3.3.0",
  "passlib[bcrypt]==1.7.4",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.2",
  "httpx==0.27.0",
  "ruff==0.6.9",
  "black==24.8.0",
  "python-dotenv==1.0.1",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP"]
